+++ test/libsolidity/StandardCompiler.cpp
@@ -903,6 +903,38 @@ BOOST_AUTO_TEST_CASE(library_linking)
 	expectLinkReferences(contractResult, {{"library2.sol", {"L2"}}});
 }
 
+BOOST_AUTO_TEST_CASE(linking_yul)
+{
+	char const* input = R"(
+	{
+		"language": "Yul",
+		"settings": {
+			"libraries": {
+				"fileB": {
+					"L": "0x4200000000000000000000000000000000000001"
+				}
+			},
+			"outputSelection": {
+				"fileA": {
+					"*": [
+						"evm.bytecode.linkReferences"
+					]
+				}
+			}
+		},
+		"sources": {
+			"fileA": {
+				"content": "object \"a\" { code { let addr := linkersymbol(\"fileB:L\") } }"
+			}
+		}
+	}
+	)";
+	Json::Value result = compile(input);
+	BOOST_TEST(containsAtMostWarnings(result));
+	Json::Value contractResult = getContractResult(result, "fileA", "a");
+	expectLinkReferences(contractResult, {});
+}
+
 BOOST_AUTO_TEST_CASE(linking_yul_empty_link_reference)
 {
 	char const* input = R"(
@@ -932,7 +964,7 @@ BOOST_AUTO_TEST_CASE(linking_yul_empty_link_reference)
 	Json::Value result = compile(input);
 	BOOST_TEST(containsAtMostWarnings(result));
 	Json::Value contractResult = getContractResult(result, "fileA", "a");
-	expectLinkReferences(contractResult, {});
+	expectLinkReferences(contractResult, {{"", {""}}});
 }
 
 BOOST_AUTO_TEST_CASE(linking_yul_no_filename_in_link_reference)
@@ -964,7 +996,39 @@ BOOST_AUTO_TEST_CASE(linking_yul_no_filename_in_link_reference)
 	Json::Value result = compile(input);
 	BOOST_TEST(containsAtMostWarnings(result));
 	Json::Value contractResult = getContractResult(result, "fileA", "a");
-	expectLinkReferences(contractResult, {});
+	expectLinkReferences(contractResult, {{"", {"L"}}});
+}
+
+BOOST_AUTO_TEST_CASE(linking_yul_same_library_name_different_files)
+{
+	char const* input = R"(
+	{
+		"language": "Yul",
+		"settings": {
+			"libraries": {
+				"fileB": {
+					"L": "0x4200000000000000000000000000000000000001"
+				}
+			},
+			"outputSelection": {
+				"fileA": {
+					"*": [
+						"evm.bytecode.linkReferences"
+					]
+				}
+			}
+		},
+		"sources": {
+			"fileA": {
+				"content": "object \"a\" { code { let addr := linkersymbol(\"fileC:L\") } }"
+			}
+		}
+	}
+	)";
+	Json::Value result = compile(input);
+	BOOST_TEST(containsAtMostWarnings(result));
+	Json::Value contractResult = getContractResult(result, "fileA", "a");
+	expectLinkReferences(contractResult, {{"fileC", {"L"}}});
 }
 
 BOOST_AUTO_TEST_CASE(evm_version)